+++ docs/scripts/routes/misc.js
@@ -1,7 +1,6 @@
 const Event = () => import('@/views/plugins/event');
 const Ripple = () => import('@/views/directives/ripple');
 const Elevation = () => import('@/views/directives/elevation');
-const TextDivider = () => import('@/views/components/text-divider');
 const Anchor = () => import('@/views/directives/anchor');
 
 export default [
@@ -19,11 +18,6 @@ export default [
     path: 'elevation',
     name: 'misc.elevation',
     component: Elevation
-  },
-  {
-    path: 'divider',
-    name: 'misc.text-divider',
-    component: TextDivider
   },
   {
     path: 'anchor',